// api/models/Department.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Department = {
    departmentID: string;
    name: string;
    description?: string;
    email?: string;
    isActive?: boolean;
    isVoided?: boolean;
    createdBy?: string;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};